truetandem/e-QIP-prototype#393 Refactored components out of HistoryCollection in preparation to reutilization in individual tracks
